Add: inorder, preorder, and postorder prints.
In-order traversal displays the values in the BST in sorted order.
Pre-order traversal can be used to copy a BST as a given node can be
created and its children would then be attached.
Post-order traversal can be used to free an entire BST as parent elements
are freed last, allowing no nodes to be abandoned.